refact analytics
