+++ pkg/analytics/algorithm/epsilongreedy/epsilongreedy.go
@@ -11,17 +11,30 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
-package analytics
+package epsilongreedy
 
 import (
-	"github.com/go-logr/logr"
+	"github.com/iter8-tools/iter8-controller/pkg/analytics/algorithm"
+	"github.com/iter8-tools/iter8-controller/pkg/analytics/api"
 	iter8v1alpha1 "github.com/iter8-tools/iter8-controller/pkg/apis/iter8/v1alpha1"
 )
 
-// AnalyticsService ...
-type AnalyticsService interface {
-	MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interface{}) (interface{}, error)
-	Invoke(log logr.Logger, endpoint string, payload interface{}, path string) (*Response, error)
-	GetPath() string
+const (
+	Strategy string = "epsilon_t_greedy"
+)
+
+var _ algorithm.Interface = Impl{}
+
+type Impl struct {
+}
+
+func (i Impl) SupplementSuccessCriteria(specSC iter8v1alpha1.SuccessCriterion, sc api.SuccessCriterion) (api.SuccessCriterion, error) {
+	return sc, nil
+}
+func (i Impl) SupplementTrafficControl(instance *iter8v1alpha1.Experiment, tc api.TrafficControl) api.TrafficControl {
+	return tc
+}
+
+func (i Impl) GetPath() string {
+	return api.AnalyticsAPIPath + Strategy
 }
